flat_test now shows good agreement
Actual and corrected rates now agree well in flat_test.m
The problem was that the dark frames were not properly corrected for thresholding and coincidence loss. This has been fixed.
Still, as the threshold factor is increased, there is a slight trend upward that is likely due to the  approximations inherent in the rand EM generator when there are more than 2 counts.